break up state slice factories
