+++ gui/src/components/tab/Settings.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from 'react-i18next';
 import { useStore } from 'zustand';
 
 import { getApiRoot } from '../../config.js';
-import { ConfigContext, StateContext, STATE_KEY } from '../../state.js';
+import { ConfigContext, StateContext, STATE_KEY } from '../../state/full.js';
 import { getTheme } from '../utils.js';
 import { NumericField } from '../input/NumericField.js';
