+++ gui/src/components/tab/Upscale.tsx
@@ -8,7 +8,8 @@ import { useStore } from 'zustand';
 import { shallow } from 'zustand/shallow';
 
 import { IMAGE_FILTER } from '../../config.js';
-import { ClientContext, OnnxState, StateContext, TabState } from '../../state.js';
+import { ClientContext, OnnxState, StateContext } from '../../state/full.js';
+import { TabState } from '../../state/types.js';
 import { HighresParams, ModelParams, UpscaleParams, UpscaleReqParams } from '../../types/params.js';
 import { Profiles } from '../Profiles.js';
 import { HighresControl } from '../control/HighresControl.js';
