+++ gui/src/state/default.ts
@@ -1,7 +1,7 @@
 import {
   BaseImgParams,
 } from '../types/params.js';
-import { TabState, Theme } from './types.js';
+import { Slice, TabState, Theme } from './types.js';
 
 export interface DefaultSlice {
   defaults: TabState<BaseImgParams>;
@@ -9,4 +9,26 @@ export interface DefaultSlice {
 
   setDefaults(param: Partial<BaseImgParams>): void;
   setTheme(theme: Theme): void;
+}
+
+export function createDefaultSlice<TState extends DefaultSlice>(defaultParams: Required<BaseImgParams>): Slice<TState, DefaultSlice> {
+  return (set) => ({
+    defaults: {
+      ...defaultParams,
+    },
+    theme: '',
+    setDefaults(params) {
+      set((prev) => ({
+        defaults: {
+          ...prev.defaults,
+          ...params,
+        }
+      } as Partial<TState>));
+    },
+    setTheme(theme) {
+      set((prev) => ({
+        theme,
+      } as Partial<TState>));
+    }
+  });
 }
