+++ gui/src/state/profile.ts
@@ -1,5 +1,6 @@
 import { Maybe } from '@apextoaster/js-utils';
 import { BaseImgParams, HighresParams, Txt2ImgParams, UpscaleParams } from '../types/params.js';
+import { Slice } from './types.js';
 
 export interface ProfileItem {
   name: string;
@@ -14,4 +15,38 @@ export interface ProfileSlice {
   removeProfile(profileName: string): void;
 
   saveProfile(profile: ProfileItem): void;
+}
+
+export function createProfileSlice<TState extends ProfileSlice>(): Slice<TState, ProfileSlice> {
+  return (set) => ({
+    profiles: [],
+    saveProfile(profile: ProfileItem) {
+      set((prev) => {
+        const profiles = [...prev.profiles];
+        const idx = profiles.findIndex((it) => it.name === profile.name);
+        if (idx >= 0) {
+          profiles[idx] = profile;
+        } else {
+          profiles.push(profile);
+        }
+        return {
+          ...prev,
+          profiles,
+        };
+      });
+    },
+    removeProfile(profileName: string) {
+      set((prev) => {
+        const profiles = [...prev.profiles];
+        const idx = profiles.findIndex((it) => it.name === profileName);
+        if (idx >= 0) {
+          profiles.splice(idx, 1);
+        }
+        return {
+          ...prev,
+          profiles,
+        };
+      });
+    }
+  });
 }
