+++ gui/src/state/upscale.ts
@@ -1,10 +1,11 @@
 import {
+  BaseImgParams,
   HighresParams,
   ModelParams,
   UpscaleParams,
   UpscaleReqParams,
 } from '../types/params.js';
-import { TabState } from './types.js';
+import { Slice, TabState } from './types.js';
 
 export interface UpscaleSlice {
   upscale: TabState<UpscaleReqParams>;
@@ -18,4 +19,69 @@ export interface UpscaleSlice {
   setUpscaleHighres(params: Partial<HighresParams>): void;
   setUpscaleModel(params: Partial<ModelParams>): void;
   setUpscaleUpscale(params: Partial<UpscaleParams>): void;
+}
+
+export function createUpscaleSlice<TState extends UpscaleSlice>(
+  defaultParams: Required<BaseImgParams>,
+  defaultHighres: HighresParams,
+  defaultModel: ModelParams,
+  defaultUpscale: UpscaleParams,
+): Slice<TState, UpscaleSlice> {
+  return (set) => ({
+    upscale: {
+      ...defaultParams,
+      // eslint-disable-next-line no-null/no-null
+      source: null,
+    },
+    upscaleHighres: {
+      ...defaultHighres,
+    },
+    upscaleModel: {
+      ...defaultModel,
+    },
+    upscaleUpscale: {
+      ...defaultUpscale,
+    },
+    resetUpscale() {
+      set({
+        upscale: {
+          ...defaultParams,
+          // eslint-disable-next-line no-null/no-null
+          source: null,
+        },
+      } as Partial<TState>);
+    },
+    setUpscale(source) {
+      set((prev) => ({
+        upscale: {
+          ...prev.upscale,
+          ...source,
+        },
+      } as Partial<TState>));
+    },
+    setUpscaleHighres(params) {
+      set((prev) => ({
+        upscaleHighres: {
+          ...prev.upscaleHighres,
+          ...params,
+        },
+      } as Partial<TState>));
+    },
+    setUpscaleModel(params) {
+      set((prev) => ({
+        upscaleModel: {
+          ...prev.upscaleModel,
+          ...defaultModel,
+        },
+      } as Partial<TState>));
+    },
+    setUpscaleUpscale(params) {
+      set((prev) => ({
+        upscaleUpscale: {
+          ...prev.upscaleUpscale,
+          ...params,
+        },
+      } as Partial<TState>));
+    },
+  });
 }
